requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.13
djangorestframework==3.15.2
idna==3.7
psycopg2==2.9.9
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
